backend/requirements.txt
blinker==1.8.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
